requirements.txt
aiogram==3.5.0
fastapi==0.114.2
uvicorn==0.30.6
SQLAlchemy==2.0.34
aiosqlite==0.20.0
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.2
openai==1.41.1
pydantic==2.7.4

